Rename renewables_clusters.py to resource_clusters.py
